+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Master
+
+- Add `Search` component
+- [Breaking change] Remove `txt-bold` and `txt-s` from `LevelIndicator` component to make it more flexible.
+
 ## 0.8.1
 
 - [Patch] Use babel.config.js and update build-package.js to output components to directories during `npm run build`.(#118)(https://github.com/mapbox/dr-ui/pull/118)